Will/metadata init update (#781)
* make user build work

* wip

* wip

* build

* feat: set failure on load fail

* clippy

+++ crates/brontes-database/brontes-db/src/libmdbx/types.rs
@@ -54,5 +54,5 @@ where
     const INIT_CHUNK_SIZE: Option<usize>;
     const INIT_QUERY: Option<&'static str>;
     const HTTP_ENDPOINT: Option<&'static str>;
-    const INIT_FLAG: Option<u8>;
+    const INIT_FLAG: Option<u16>;
 }
